Add NYC taxis track to nightly benchmarks
With this commit we add the NYC taxis track to the nightly
benchmark suite. The benchmark will be run and also all reports
are generated but the page is not yet added to the main menu so
we have a chance to review the results first before making them
really public.

Relates #2