Disable query benchmarking for PMC/defaults
With this commit we disable query benchmarking for the `pmc` track and
the `defaults` car because scroll queries take up a lot of time recently
and we don't show it in the graphs (because we use the results from the
`4gheap` car for that). This lets our nightly benchmark take up a lot of
time.

In parallel we will investigate why scroll is so slow recently and will
reenable query benchmarking again after we found out.

Relates #19