Use Kibana time series visualizations for benchmark results
With this commit, night_rally uses Kibana for visualizing nightly and
release benchmark results. We also add an administration tool e.g. for
adding annotations.

While we add support for these new visualizations now, we are in a
transitory phase and will still populate and use the current charts for
the time being. After a grace period, we will then switch to using
Kibana exclusively.

Relates #23