Start with maximum configuration
With this commit we ensure the nightly benchmarks start with the
"maximum" configuration, which is at the moment a three node benchmark.
This is necessary because night_rally enforces a build on the first step
but skips it afterwards. We could (and maybe also will) implement a
detection logic whether a build is needed but for now the pragmatic
choice is just to start with the maximum configuration

Relates #22