Reset benchmark environment to clean state
With this commit we reset the benchmark environment to a clean state
before running each combination. Previously we only reset the system to
a clean state at the beginning of the benchmark suite execution.

From now on we are giving all benchmarks the same starting conditions
and thus are avoiding ordering bias.