Don't write CSV race report anymore
With this commit we will not write the results as CSV files anymore. A
long time ago we needed this because the result charts were based on
these reports but since we moved to Kibana, all metrics are kept and
read there. Besides, Rally also creates results in a structured
race.json file anyway so the result reports are not necessary anymore.