Explicitly specify numbers of shards for nyc_taxis
With this commit we do not only specify the number of replicas for
nyc_taxis but also the number of shards (for the three node case). The
reason is that previously the number of shards has been specified as
part of the challenge and thus we have automatically overridden it by
providing the track parameter `index_settings`. However, we want to move
the number of shards from the challenge to the index definition and
there we need to be explicit (there is no override) thus we need to set
the number of shards explicitly to 5.