Refactor chart configuration
With this commit we remove the parameter "default-combination" and
replace it with a "charts" array. Previously, "default-combination"
indicated that we want to generate all charts whereas if it was not
present, we only created indexing charts. With the "charts" array it is
now clear which charts will be generated by Rally.

We also rename the "combinations" array to "configurations" because this
name is more appropriate.

Relates #77