Specify any params structured
With this commit we get rid of all string-based representation of more complex
parameters. This affects track-params and car-params specifically which will
now be passed to Rally as JSON string in any case. This allows to merge those
parameters in more complex scenarios where we need to source them from multiple
places.

Closes #82
Relates #83