Add query charts for trial license
With this commit we amend the existing race configuration so the chart
generator also creates query charts for the trial license (which we use
to run with x-pack). We also change the dashboard URLs as new dashboards
and charts have already been generated. The new web page is already
live.

Relates #106
Relates elastic/rally#684