Apply setup user-tag logic consistently
With this commit we always take all parameters into account when
determining the `setup` user-tag. Previously, we have always used "bare"
as value regardless of the specific setup. This means that the query
that attempts to deactivate outdated results will never match any
document (because they are copied to the release environment with the
proper setup tag). Queries that rely on this logic may produce invalid
results as a consequence.

Relates #133