Enable GC logs for geonames (#198)
Since February 7 we noticed an increase in the 100th percentile service
time of the `country_agg_cached` query in geonames when the trial
license is enabled. In order to dig whether this is related to JVM
activity (GC, safepoints) we temporarily enable GC logs. This is
expected to be low-impact on performance but allows us to dig deeper.

Relates #181