Trigger heapdump on request errors in nyc_taxis (#213)
With this commit we expose the `on-error` command line argument in
night-rally's race configurations. We configure all G1-related
configurations for nyc_taxis to abort on error and add a heapdump
telemetry device so we get a heapdump in this case for further
inspection. Finally, we ensure that the build won't fail because of an
(expected) error.

Relates #204