Enable segment stats for G1 benchmarks (#250)
With this commit we enable the new segment stats telemetry device
introduced by elastic/rally#988. The primary purpose is to aid the
investigation in #204. Please note that this telemetry device will
capture data via an API call and will thus write the log file on the
coordinator machine and *not* on the target machine(s).

Relates #204
Relates elastic/rally#988