Handle ports more resiliently (#254)
With this commit we switch to Elasticsearch's standard ports (9200 for
HTTP traffic and 9300 for the transport port). Originally, we have
chosen port 39200 for HTTP traffic in order to avoid conflicts with any
other Elasticsearch processes running on the machine(s). However, having
an unexpected Elasticsearch process running during a benchmark is a flaw
in benchmark methodology and we should thus not guard against it. Also,
by lowering the port we avoid the usual ephemeral port range (on our
machines 32768 to 60999, see `/proc/sys/net/ipv4/ip_local_port_range`)
and reduce the risk of unintended port conflicts. Furhtermore, we amend
night-rally's check for free ports on the target machines to also check
the transport port.