Keep running the same tasks in nyc_taxis (#278)
elastic/rally-tracks#127 adds
`force-merge/refresh/wait-until-merges-finish` tasks for ad-hoc
investigations related to the cost of force-merge.

In this commit we explicitly skip the first two tasks to ensure nightlies keep
running the same tasks as they used to. Not running `wait-until-merges-finish`
was a bug and we want it executed on nightlies.