Add benchmark configuration with JFR enabled (#317)
With this commit we add a new configuration
`geopoint-append-4g-1node-with-jfr` which is intended to measure the
overhead of enabling Java flight recorder. Oracle claims that enabling
Java flight recorder with the `default` recording profile has at most an
overhead of 2% (see [1], page 6) but we'd like to measure this
ourselves. There is no need to run this configuration for all workloads.
Instead, we chose `geopoint` because it is a very sensitive workload.

Relates elastic/petete#7

[1] https://www.oracle.com/technetwork/java/javaseproducts/mission-control/java-mission-control-wp-2008279.pdf