Don't purge CDN cache after page update (#329)
With this commit we don't purge the CDN cache anymore after a new change
to the benchmarks page has been deployed but add a clarifying note about
the CDN cache expiration time. We are dropping cache purge after infra
has moved from Fastly to Google's CDN where they have configured a cache
expiry of 10 minutes. Given that the timeout is so small, we don't push
changes very often and purging the cache on Google's CDN is a bit more
involved, our team has decided that we just let the stale cache entry
expire.

Closes #322