Add dedicated page for incubating benchmarks (#347)
With this commit we add a new benchmark page for incubating benchmarks.
The intention is to remove this page after a while again so we have just
duplicated the existing index page. If we require such a page for a
longer time period we should make an attempt to unify them.

As the first benchmark we move EQL there. While the development team is
interested in benchmarks, we should not publish them yet along the other
benchmarks.

Relates #345
Relates #346