Improve filters for night-rally-admin (#354)
In earlier iterations of the nightlies we did not make effective use of
the user tags feature to identify race configurations. However, we add
user tags to identify the license with which we run benchmarks and we
also use a descriptive name for each configuration we run. Contrary, the
specific car and challenge are not as useful to filter races.

With this commit we therefore adapt the available filters in the
`night-rally-admin` tool to drop the `--car` and `--challenge` filters
and introduce two new filter flags `--license` and `--name`. With this
new filters it is easier to narrow down the list of relevant races to
raise an investigation issue.