Update Vagrantfile commands to set pyenv shims (#462)
With this commit we ensure that `pyenv` is correctly installed and configured. The `pyenv init -` command no longer adds the shims to `PATH`, and instead must now be done with a new command: `pyenv init --path`. Without this, provisioning the VMs with vagrant up will fail on the `install_rally_source` function. Additionally, the installation from git checkout has been simplified to a single line along with ensuring the ordering of the `pyenv` commands.

Relates to https://github.com/elastic/elasticsearch-benchmarks/issues/444