Generate annotation ids (#533)
Generate annotation IDs client-side.

Previously, when indexing a document representing a new annotation, we relied on
Elasticsearch to generate the document's ID. Elasticsearch generates IDs by
base64-encoding a UUID in URL-safe mode. This can result in IDs with leading
hyphens, which argparse cannot parse. We now instead generate a UUID ourselves
and use it as the document's ID.

This commit also adds a config file for `isort`, which was run on `night_rally/admin.py`.

Co-authored-by: Quentin Pradet <quentin.pradet@elastic.co>