Re-introduce aggs 1g (#670)
In May we decided to remove this configuration from night-rally after seeing multiple failures. We are now unable to reproduce the failures for the configuration in manual testing, so I'd propose we add it back to nightlies, with a delay to adding charts back for the challenge. Once the challenge proves itself to be stable, or else reproduces the issue and is thoroughly investigated, we will take further steps.

Relates to elastic/elasticsearch#86531 (and its children)